requirements.txt
requests
Flask
gevent>=1.1.1
stups-tokens
stups-pierone
